+++ Magent/Views/Settings/SettingsProjectsViewController.swift
@@ -21,7 +21,7 @@ final class SettingsProjectsViewController: NSViewController {
     var repoPathLabel: NSTextField!
     var worktreesPathLabel: NSTextField!
     var defaultBranchField: NSTextField!
-    var localFileSyncPathsTextView: NSTextView!
+    var localFileSyncEntriesStack: NSStackView!
     var agentTypePopup: NSPopUpButton!
     var terminalInjectionTextView: NSTextView!
     var preAgentInjectionTextView: NSTextView!